@@ -33,9 +33,9 @@
     "babel-preset-env": "^1.1.8",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
-    "enzyme": "^2.7.1",
-    "eslint": "^3.13.1",
-    "eslint-plugin-react": "^6.9.0",
+    "enzyme": "^2.8.2",
+    "eslint": "^3.19.0",
+    "eslint-plugin-react": "^6.10.3",
     "jasmine": "^2.5.3",
     "jasmine-core": "^2.5.2",
     "json-loader": "^0.5.4",
@@ -44,10 +44,11 @@
     "karma-jasmine": "^1.1.0",
     "karma-webpack": "^2.0.2",
     "react-addons-test-utils": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "webpack": "^1.14.0"
+    "react-dom": "^15.5.4",
+    "webpack": "^1.15.0"
   },
   "dependencies": {
-    "react": "^15.4.2"
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4"
   }
 }
